java language bindings 3.1-rc3
