+++ models/java/gbfs-java-model/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.mobilitydata</groupId>
     <artifactId>gbfs-java-model</artifactId>
-    <version>1.0.12</version>
+    <version>1.0.13</version>
 
     <name>gbfs-java-model</name>
 
@@ -257,7 +257,7 @@
                         </configuration>
                     </execution>
                     <execution>
-                        <id>generate-v3.1-RC2-sources</id>
+                        <id>generate-v3.1-RC3-sources</id>
                         <goals>
                             <goal>generate</goal>
                         </goals>
@@ -270,8 +270,8 @@
                             <serializable>true</serializable>
                             <initializeCollections>false</initializeCollections>
                             <generateBuilders>true</generateBuilders>
-                            <sourceDirectory>${basedir}/resources/v3.1-RC2</sourceDirectory>
-                            <targetPackage>org.mobilitydata.gbfs.v3_1_RC2</targetPackage>
+                            <sourceDirectory>${basedir}/resources/v3.1-RC3</sourceDirectory>
+                            <targetPackage>org.mobilitydata.gbfs.v3_1_RC3</targetPackage>
                             <customRuleFactory>org.mobilitydata.gbfs.GeofencingRuleFactory</customRuleFactory>
                         </configuration>
                     </execution>
