--- models/java/gbfs-java-model/src/test/java/org/mobilitydata/gbfs/v3_1_RC2/gbfs_versions/GBFSVersionsTest.java
@@ -1,12 +1,0 @@
-package org.mobilitydata.gbfs.v3_1_RC2.gbfs_versions;
-
-import org.mobilitydata.gbfs.TestBase;
-
-import org.junit.jupiter.api.Test;
-
-class GBFSVersionsTest extends TestBase {
-    @Test
-    void testUnmarshal() {
-        assertUnmarshalDoesNotThrow("v3_1_RC2/gbfs_versions.json", GBFSGbfsVersions.class);
-    }
-}
