--- models/java/gbfs-java-model/src/test/java/org/mobilitydata/gbfs/v3_1_RC2/system_alerts/SystemAlertsTest.java
@@ -1,11 +1,0 @@
-package org.mobilitydata.gbfs.v3_1_RC2.system_alerts;
-
-import org.mobilitydata.gbfs.TestBase;
-import org.junit.jupiter.api.Test;
-
-class SystemAlertsTest extends TestBase {
-    @Test
-    void testUnmarshal() {
-        assertUnmarshalDoesNotThrow("v3_1_RC2/system_alerts.json", GBFSSystemAlerts.class);
-    }
-}
