+++ models/java/gbfs-java-model/src/test/java/org/mobilitydata/gbfs/v3_1_RC3/manifest/ManifestTest.java
@@ -1,0 +1,12 @@
+package org.mobilitydata.gbfs.v3_1_RC3.manifest;
+
+import org.mobilitydata.gbfs.TestBase;
+
+import org.junit.jupiter.api.Test;
+
+class ManifestTest extends TestBase {
+    @Test
+    void testUnmarshal() {
+        assertUnmarshalDoesNotThrow("v3_1_RC3/manifest.json", GBFSManifest.class);
+    }
+}
