+++ models/java/gbfs-java-model/src/test/java/org/mobilitydata/gbfs/v3_1_RC3/system_information/SystemInformationTest.java
@@ -1,0 +1,11 @@
+package org.mobilitydata.gbfs.v3_1_RC3.system_information;
+
+import org.mobilitydata.gbfs.TestBase;
+import org.junit.jupiter.api.Test;
+
+class SystemInformationTest extends TestBase {
+    @Test
+    void testUnmarshal() {
+        assertUnmarshalDoesNotThrow("v3_1_RC3/system_information.json", GBFSSystemInformation.class);
+    }
+}
