+++ models/java/gbfs-java-model/src/test/java/org/mobilitydata/gbfs/v3_1_RC3/system_pricing_plans/SystemPricingPlansTest.java
@@ -1,0 +1,11 @@
+package org.mobilitydata.gbfs.v3_1_RC3.system_pricing_plans;
+
+import org.mobilitydata.gbfs.TestBase;
+import org.junit.jupiter.api.Test;
+
+class SystemPricingPlansTest extends TestBase {
+    @Test
+    void testUnmarshal() {
+        assertUnmarshalDoesNotThrow("v3_1_RC3/system_pricing_plans.json", GBFSSystemPricingPlans.class);
+    }
+}
