+++ models/java/gbfs-java-model/src/test/java/org/mobilitydata/gbfs/v3_1_RC3/vehicle_availability/VehicleAvailabilityTest.java
@@ -1,0 +1,11 @@
+package org.mobilitydata.gbfs.v3_1_RC3.vehicle_availability;
+
+import org.mobilitydata.gbfs.TestBase;
+import org.junit.jupiter.api.Test;
+
+class VehicleAvailabilityTest extends TestBase {
+	@Test
+	void testUnmarshal() {
+		assertUnmarshalDoesNotThrow("v3_1_RC3/vehicle_availability.json", GBFSVehicleAvailability.class);
+	}
+}
