+++ models/java/gbfs-java-model/src/test/java/org/mobilitydata/gbfs/v3_1_RC3/vehicle_types/VehicleTypesTest.java
@@ -1,0 +1,11 @@
+package org.mobilitydata.gbfs.v3_1_RC3.vehicle_types;
+
+import org.mobilitydata.gbfs.TestBase;
+import org.junit.jupiter.api.Test;
+
+class VehicleTypesTest extends TestBase {
+    @Test
+    void testUnmarshal() {
+        assertUnmarshalDoesNotThrow("v3_1_RC3/vehicle_types.json", GBFSVehicleTypes.class);
+    }
+}
