--- models/java/gbfs-java-model/src/test/resources/v3_1_RC2/gbfs.json
@@ -1,17 +1,0 @@
-{
-    "last_updated": "2023-07-17T13:34:13+02:00",
-    "ttl": 0,
-    "version": "3.1-RC2",
-    "data": {
-      "feeds": [
-        {
-          "name": "system_information",
-          "url": "https://www.example.com/gbfs/1/system_information"
-        },
-        {
-          "name": "station_information",
-          "url": "https://www.example.com/gbfs/1/station_information"
-        }
-      ]
-    }
-  }
