--- models/java/gbfs-java-model/src/test/resources/v3_1_RC2/gbfs_versions.json
@@ -1,17 +1,0 @@
-{
-    "last_updated": "2023-07-17T13:34:13+02:00",
-    "ttl": 0,
-    "version": "3.1-RC2",
-    "data": {
-      "versions": [
-        {
-          "version": "2.0",
-          "url": "https://www.example.com/gbfs/2/gbfs"
-        },
-        {
-          "version": "3.1-RC2",
-          "url": "https://www.example.com/gbfs/3.1-RC2/gbfs"
-        }
-      ]
-    }
-  }
