--- models/java/gbfs-java-model/src/test/resources/v3_1_RC2/system_regions.json
@@ -1,45 +1,0 @@
-{
-    "last_updated": "2023-07-17T13:34:13+02:00",
-    "ttl": 86400,
-    "version": "3.1-RC2",
-    "data": {
-      "regions": [
-        {
-          "name": [
-            {
-              "text": "North",
-              "language": "en"
-            }
-          ],
-          "region_id": "3"
-        },
-        {
-          "name": [
-            {
-              "text": "East",
-              "language": "en"
-            }
-          ],
-          "region_id": "4"
-        },
-        {
-          "name": [
-            {
-              "text": "South",
-              "language": "en"
-            }
-          ],
-          "region_id": "5"
-        },
-        {
-          "name": [
-            {
-              "text": "West",
-              "language": "en"
-            }
-          ],
-          "region_id": "6"
-        }
-      ]
-    }
-  }
