+++ models/java/gbfs-java-model/src/test/resources/v3_1_RC3/gbfs_versions.json
@@ -1,0 +1,17 @@
+{
+    "last_updated": "2023-07-17T13:34:13+02:00",
+    "ttl": 0,
+    "version": "3.1-RC3",
+    "data": {
+      "versions": [
+        {
+          "version": "2.0",
+          "url": "https://www.example.com/gbfs/2/gbfs"
+        },
+        {
+          "version": "3.1-RC3",
+          "url": "https://www.example.com/gbfs/3.1-RC3/gbfs"
+        }
+      ]
+    }
+  }
