+++ models/java/gbfs-java-model/src/test/resources/v3_1_RC3/geofencing_zones.json
@@ -1,0 +1,74 @@
+{
+    "last_updated": "2023-07-17T13:34:13+02:00",
+    "ttl": 60,
+    "version": "3.1-RC3",
+    "data": {
+      "geofencing_zones": {
+        "type": "FeatureCollection",
+        "features": [
+          {
+            "type": "Feature",
+            "geometry": {
+              "type": "MultiPolygon",
+              "coordinates": [
+                [
+                  [
+                    [
+                      -122.578067,
+                      45.562982
+                    ],
+                    [
+                      -122.661838,
+                      45.562741
+                    ],
+                    [
+                      -122.661151,
+                      45.504542
+                    ],
+                    [
+                      -122.578926,
+                      45.5046625
+                    ],
+                    [
+                      -122.578067,
+                      45.562982
+                    ]
+                  ]
+                ]
+              ]
+            },
+            "properties": {
+              "name": [
+                {
+                  "text": "NE 24th/NE Knott",
+                  "language": "en"
+                }
+              ],
+              "start": "2023-07-17T13:34:13+02:00",
+              "end": "2024-07-18T13:34:13+02:00",
+              "rules": [
+                {
+                  "vehicle_type_ids": [
+                    "moped1",
+                    "car1"
+                  ],
+                  "ride_start_allowed": true,
+                  "ride_end_allowed": true,
+                  "maximum_speed_kph": 10,
+                  "station_parking": true,
+                  "ride_through_allowed": false 
+                }
+              ]
+            }
+          }
+        ]
+      },
+      "global_rules": [
+        {
+          "ride_start_allowed": false,
+          "ride_end_allowed": false,
+          "ride_through_allowed": true
+        }
+      ]
+    }
+  }
