+++ models/java/gbfs-java-model/src/test/resources/v3_1_RC3/manifest.json
@@ -1,0 +1,115 @@
+{
+    "last_updated": "2023-07-17T13:34:13+02:00",
+    "ttl":0,
+    "version": "3.1-RC3",
+    "data":{
+      "datasets":[
+        {
+          "system_id":"example_berlin",
+          "versions":[
+            {
+              "version": "2.0",
+              "url":"https://berlin.example.com/gbfs/2/gbfs"
+            },
+            {
+              "version": "3.1-RC3",
+              "url":"https://berlin.example.com/gbfs/3.1-RC3/gbfs"
+            }
+          ],
+          "area": {
+            "type": "MultiPolygon",
+            "coordinates": [
+              [
+                [
+                  [
+                    13.10821,
+                    52.58563
+                  ],
+                  [
+                    13.29743,
+                    52.67046
+                  ],
+                  [
+                    13.48451,
+                    52.6855
+                  ],
+                  [
+                    13.77993,
+                    52.43458
+                  ],
+                  [
+                    13.65355,
+                    52.33048
+                  ],
+                  [
+                    13.08165,
+                    52.38793
+                  ],
+                  [
+                    13.10821,
+                    52.58563
+                  ]
+                ]
+              ]
+            ]
+          },
+          "country_code": "DE"
+        },
+        {
+          "system_id":"example_paris",
+          "versions":[
+            {
+              "version": "2.0",
+              "url":"https://paris.example.com/gbfs/2/gbfs"
+            },
+            {
+              "version": "3.1-RC3",
+              "url":"https://paris.example.com/gbfs/3.1-RC3/gbfs"
+            }
+          ],
+          "area": {
+            "type": "MultiPolygon",
+            "coordinates": [
+              [
+                [
+                  [
+                    2.14306,
+                    48.89971
+                  ],
+                  [
+                    2.36707,
+                    48.99455
+                  ],
+                  [
+                    2.60219,
+                    49.01987
+                  ],
+                  [
+                    2.615,
+                    48.69025
+                  ],
+                  [
+                    2.52167,
+                    48.6867
+                  ],
+                  [
+                    2.26838,
+                    48.73275
+                  ],
+                  [
+                    2.13103,
+                    48.80833
+                  ],
+                  [
+                    2.14306,
+                    48.89971
+                  ]
+                ]
+              ]
+            ]
+          },
+          "country_code": "FR"
+        }
+      ]
+    }
+  }
