+++ models/java/gbfs-java-model/src/test/resources/v3_1_RC3/vehicle_status.json
@@ -1,0 +1,32 @@
+{
+  "last_updated": "2023-07-17T13:34:13+02:00",
+  "ttl":0,
+  "version": "3.1-RC3",
+  "data":{
+    "vehicles":[
+      {
+        "vehicle_id":"973a5c94-c288-4a2b-afa6-de8aeb6ae2e5",
+        "last_reported": "2023-07-17T13:34:13+02:00",
+        "lat":12.345678,
+        "lon":56.789012,
+        "is_reserved":false,
+        "is_disabled":false,
+        "vehicle_type_id":"abc123",
+        "rental_uris": {
+          "android": "https://www.example.com/app?vehicle_id=973a5c94-c288-4a2b-afa6-de8aeb6ae2e5&platform=android&",
+          "ios": "https://www.example.com/app?vehicle_id=973a5c94-c288-4a2b-afa6-de8aeb6ae2e5&platform=ios"
+        }
+      },
+      {
+        "vehicle_id":"987fd100-b822-4347-86a4-b3eef8ca8b53",
+        "last_reported": "2023-07-17T13:34:13+02:00",
+        "is_reserved":false,
+        "is_disabled":false,
+        "vehicle_type_id":"def456",
+        "current_range_meters":6543.0,
+        "station_id":"86",
+        "pricing_plan_id":"plan3"
+      }
+    ]
+  }
+}
