+++ models/java/gbfs-java-model/src/test/resources/v3_1_RC3/vehicle_types.json
@@ -1,0 +1,143 @@
+{
+  "last_updated": "2023-07-17T13:34:13+02:00",
+  "ttl": 0,
+  "version": "3.1-RC3",
+  "data": {
+    "vehicle_types": [
+      {
+        "vehicle_type_id": "abc123",
+        "form_factor": "bicycle",
+        "propulsion_type": "human",
+        "name": [
+          {
+            "text": "Example Basic Bike",
+            "language": "en"
+          }
+        ],
+        "wheel_count": 2,
+        "default_reserve_time": 30,
+        "return_constraint": "any_station",
+        "min_age": 14,
+        "vehicle_assets": {
+          "icon_url": "https://www.example.com/assets/icon_bicycle.svg",
+          "icon_url_dark": "https://www.example.com/assets/icon_bicycle_dark.svg",
+          "icon_last_modified": "2021-06-15"
+        },
+        "default_pricing_plan_id": "bike_plan_1",
+        "pricing_plan_ids": [
+          "bike_plan_1",
+          "bike_plan_2",
+          "bike_plan_3"
+        ]
+      },
+      {
+        "vehicle_type_id": "cargo123",
+        "form_factor": "cargo_bicycle",
+        "propulsion_type": "human",
+        "name": [
+          {
+            "text": "Example Cargo Bike",
+            "language": "en"
+          }
+        ],
+        "description": [
+          {
+            "text": "Extra comfortable seat with additional suspension.\n\nPlease be aware of the cargo box lock: you need to press it down before pulling it up again!",
+            "language": "en"
+          }
+        ],            
+        "wheel_count": 3,
+        "default_reserve_time": 30,
+        "return_constraint": "roundtrip_station",
+        "min_age": 18,
+        "vehicle_assets": {
+          "icon_url": "https://www.example.com/assets/icon_cargobicycle.svg",
+          "icon_url_dark": "https://www.example.com/assets/icon_cargobicycle_dark.svg",
+          "icon_last_modified": "2021-06-15"
+        },
+        "default_pricing_plan_id": "cargo_plan_1",
+        "pricing_plan_ids": [
+          "cargo_plan_1",
+          "cargo_plan_2",
+          "cargo_plan_3"
+        ]
+      },
+      {
+        "vehicle_type_id": "def456",
+        "form_factor": "scooter_standing",
+        "propulsion_type": "electric",
+        "name": [
+          {
+            "text": "Example E-scooter V2",
+            "language": "en"
+          }
+        ],
+        "wheel_count": 2,
+        "max_permitted_speed": 25,
+        "rated_power": 350,
+        "default_reserve_time": 30,
+        "max_range_meters": 12345,
+        "return_constraint": "free_floating",
+        "vehicle_assets": {
+          "icon_url": "https://www.example.com/assets/icon_escooter.svg",
+          "icon_url_dark": "https://www.example.com/assets/icon_escooter_dark.svg",
+          "icon_last_modified": "2021-06-15"
+        },
+        "default_pricing_plan_id": "scooter_plan_1"
+      },
+      {
+        "vehicle_type_id": "car1",
+        "form_factor": "car",
+        "rider_capacity": 5,
+        "cargo_volume_capacity": 200,
+        "propulsion_type": "combustion_diesel",
+        "eco_labels": [
+          {
+            "country_code": "FR",
+            "eco_sticker": "critair_1"
+          },
+          {
+            "country_code": "DE",
+            "eco_sticker": "euro_2"
+          }
+        ],
+        "name": [
+          {
+            "text": "Four-door Sedan",
+            "language": "en"
+          }
+        ],
+        "wheel_count": 4,
+        "default_reserve_time": 0,
+        "max_range_meters": 523992,
+        "return_constraint": "roundtrip_station",
+        "vehicle_accessories": [
+          "doors_4",
+          "automatic",
+          "cruise_control"
+        ],
+        "g_CO2_km": 120,
+        "vehicle_image": "https://www.example.com/assets/renault-clio.jpg",
+        "make": [
+          {
+            "text": "Renault",
+            "language": "en"
+          }
+        ],
+        "model": [
+          {
+            "text": "Clio",
+            "language": "en"
+          }
+        ],
+        "color": "white",
+        "vehicle_assets": {
+          "icon_url": "https://www.example.com/assets/icon_car.svg",
+          "icon_url_dark": "https://www.example.com/assets/icon_car_dark.svg",
+          "icon_last_modified": "2021-06-15"
+        },
+        "default_pricing_plan_id": "car_plan_1"
+      }
+    ]
+  }
+}
